client/requirements.txt
cryptography==2.8
pyjwt==1.7.0
paho-mqtt==1.3.1
protobuf==3.10.0
